make it easier to access all of the capture primitives
