+++ doc/releases/changelog-dev.md
@@ -158,6 +158,9 @@
   [(#6041)](https://github.com/PennyLaneAI/pennylane/pull/6041)
   [(#6064)](https://github.com/PennyLaneAI/pennylane/pull/6064)
 
+* Capture project objects with jax dependencies can now be accessed via the `pennylane.capture.primitives`
+  submodule.  This module is not automatically imported with pennylane.
+
 * `qml.for_loop` and `qml.while_loop` now fallback to standard Python control
   flow if `@qjit` is not present, allowing the same code to work with and without
   `@qjit` without any rewrites.
