+++ tests/capture/test_capture_qnode.py
@@ -21,11 +21,13 @@ from functools import partial
 import pytest
 
 import pennylane as qml
-from pennylane.capture import qnode_prim
 
 pytestmark = pytest.mark.jax
 
 jax = pytest.importorskip("jax")
+
+# must be below jax importorskip
+from pennylane.capture.primitives import qnode_prim  # pylint: disable=wrong-import-position
 
 
 @pytest.fixture(autouse=True)
